+++ opal/mca/hwloc/hwloc-internal.h
